Fix `toBest` returning first result even if it's less than the `cutOffNumber` (#222)
* Fix toBest returning a number less than the cut off number

Work in progress but I think it solves the issue at hand

If the no other results have been selected, if the first unit
converted produces a result less than 1, the code will accept that
result as the best. I don't believe this is the desired result. More
tests and discussion needs to happen.

* Update test name to be more descriptive

* Fix linting issues in tests