:recycle: Correct form of non-abbreviated kelvin scale. (#270)
As per Wikipedia (https://en.wikipedia.org/wiki/Kelvin), the Kelvin scale
is never measured in "degrees". Instead, the name is usually pluralized
(e.g. "kelvins") to refer to multiple units within this scale. While it
isn't supposed to be capitalized, common writing trends tend to capitalize
all units of temperature measurement, so this was retained from the
original version.