Rework package.json (#305)
* Rework package.json

The issue #302 recommended a tool called publint, which showed various
issues with the project's package.json. Since I'm no expert when it
comes to packaging, I think it would be very wise to follow the issue
author's advice on fixing the publint errors/warnings and also running
publint on PRs.

I think going forward, code changes should not cause errors or warning
when publint is ran. Aside from the fact that this will aid in testing
and also serve as a sanity check so that releases don't break and cause
folks a ton of headaches.

This will also fix the node-ci action as it was trying to use node
versions that haven't been released yet. I don't know why I added those
and why I never tested it. :clown:

* Remove Typecheck action

This action is not really required. The compile step will report any
issues so a second check is not really helpful and just prolongs the CI
pipeline.