Fix unit type issues
The main goal of these changes is to allow for any string to be passed
to methods like `to` and `from` while still providing enough information
that the developer's editor will still provide a list of possible units.

The problem with the old implementation is that methods like `to` and
`from` expect `TUnits`. Therefore, typescript expects the value being
passed to be a member of that type. This isn't a big deal if the value
is a string literal since typescript will infer the value's type to be
`TUnits`. However, for cases where the value is dynamic, that value's
type has to be explicitly set to `TUnits` or a type guard needs to be
used. Initially, I thought type guards were a good solution to this
problem but after some thought, I realised that the unit string would
then be checked twice. Once for the type guard and then by the method
since it needs to verify that the provided unit string exists. This is
also a pain to have to do all the time and adds a lot of boiler plate
code to any project using this lib. This commit removes the need to use
type guards by changing various method parameter types from `TUnit` to
`TUnit | (string & {})`. The reason why this works and `TUnit | string`
does not, is because the type definition for `TUnit | string` gets
simplified to just `string` since `TUnit` is within the `string` type.
I'll admit, I don't exactly know why `TUnit | (string & object)` but
what I do know is that it tricks typescript into not simplifying the
type which then allows for any member of `TUnit` or any `string`.

Why not just change everything to the type of `string`? Because I would
still like my editor to suggest the units as it helps prevent mistakes
when using literal units. This is the whole reason why someone would
want to use typescript in the first place and configure their editor to
give them feedback based on the types.

This commit will fix #217 and a few others.

Here are the other changes:

- Updated jest config to ignore `TS151001`
- Move jest config from javascript to typescript
- Update jest and ts-jest versions to the latest
- Remove `@ts-expect-error` flags from tests as they will not longer
  produce typescript errors
- Add explicit error classes and JSDocs `@throws` for methods that can
  throw errors. The goal is to make it more obvious where and what
  errors can be thrown. Unfortunately, there isn't a way to do this with
  typescript.
- The methods with updated argument types are: possibilities, list,
  describe, getUnit, toBest, to, and from