Update `toBest` to return the original value
This will update the `toBest` to require the original value/unit if a
better value is not found. The previous behaviour was to return `null`.
This will also be the case if the value is zero.

This will also add these cases to the documentation as well as add a
note stating that the *best* value is subjective and that `toBest` does
not work for all measures.

Tests were updated to reflect the changes and one test was removed
because it was testing the same thing as another.