Update Digital Units (#316)
* Update Digital Units

These changes should have happened a long time ago. Sorry that it took
this long to sort out.

Note that these changes modify the unit names for the digital measure.
Therefore, this is a breaking change if you are using the measure. The
system names were also changed to be more clear and map to their
respective standards. The bit and byte units were moved to their own
systems to avoid duplication. I think it makes a lot more sense like
this and it also makes the conversions work correctly.

These changes will add  the SI system and the IEC system. Again, byte
and bit are their own systems.

* Update digital units in README