Small performance improvement
Updated the Converter class so that it accepts a unit cache. Then
updated the getUnit method to use the cache. This will make the `to` and
`from` methods execute a bit faster.

This also fixes the index imports as not everything that should have
been exported was being exported. Also fixed the default name not
matching the docs.