Alternative number support (#351)
* Added support for alternative number objects

This is still a work in progress at this stage. However, I'm happy
enough with the results and implementation that I think it is worth
sharing to get some feedback. With that said, it is possible that the
implementation might change so use it at your own risk :)

A user that doesn't want to use another number object, eg. Decimal.js,
does not need to make any changes to their code base. The implementation
is completely backwards compatible. The existing tests prove that since
I didn't modify any of them to get this to work correctly. However,
there is one small difference that will only affect users that define
their own measures or extends the existing measures. The definitions
`to_anchor` and `ratio` properties now support being defined as a
fraction or a number. Before a `to_anchor` value would have been defined
with the expression, `1 / 1000`. The expression would be evaluated at
startup which was fine but it would still be vulnerable to floating
point errors since it wouldn't use the desired number object (like
Decimal.js) to evaluate the expression. The solution is to allow those
properties to be defined as a fraction that would be evaluated later
using the desired number object. The catch is that any custom measures
will need to be updated to fractions but only if the user cares about
the issue it's trying to solve. Using the `1 / 100` expression as the
`to_anchor` properties' value is still valid. So nothing will break
but the benefits of the new implementation will not be seen until
those measures are updated.

With these changes, the `configureMeasures` function now accepts a new
type argument and an object that implements the `Wrapper<T>` interface.
These will allow a user to use something like Decimal.js if they want.
If the wrapper object is not provided, the library will use a default
wrapper which uses `number`. See the `DefaultWrapper` and Decimal.js
tests for details on how to use the `Wrapper<T>` interface.

All methods that once returned a `number` will now return the type that
the `Wrapper<T>` implements. Therefore, if the default wrapper is used,
those methods will return a `number`. If a wrapper that use Decimal.js
is provided, those methods will return `Decimal` objects.

For compatibility with various different number object libraries,
passing strings is now supported. Though, I'm not completely sold on
the idea and still need to write test cases to make sure nothing breaks.

Additional testing is still required before I am willing/comfortable
merging these changes into the main branch.

* Fix package-lock.json and lint errors

* Updated packages with known vulnerabilities